#661 moved test method to proper file
