+++ bindings/py/tests/algorithms/temporal_memory_test.py
@@ -25,7 +25,28 @@ from htm.bindings.sdr import SDR
 from htm.algorithms import TemporalMemory as TM
 import numpy as np
 
+debugPrint = False
 
+parameters1 = {
+
+ 'sp': {'boostStrength': 3.0,
+        'columnCount': 1638,
+        'localAreaDensity': 0.04395604395604396,
+        'potentialPct': 0.85,
+        'synPermActiveInc': 0.04,
+        'synPermConnected': 0.13999999999999999,
+        'synPermInactiveDec': 0.006},
+ 'tm': {'activationThreshold': 17,
+        'cellsPerColumn': 13,
+        'initialPerm': 0.21,
+        'maxSegmentsPerCell': 128,
+        'maxSynapsesPerSegment': 64,
+        'minThreshold': 10,
+        'newSynapseCount': 32,
+        'permanenceDec': 0.1,
+        'permanenceInc': 0.1},
+}
+ 
 class TemporalMemoryBindingsTest(unittest.TestCase):
 
   def testCompute(self):
@@ -88,3 +109,73 @@ class TemporalMemoryBindingsTest(unittest.TestCase):
     tm3.loadFromFile(file)
     self.assertEqual(str(tm), str(tm3), "TemporalMemory serialization (using saveToFile/loadFromFile) failed.")
     os.remove(file)
+
+  def testPredictiveCells(self):
+    """
+    This tests that we don't get empty predicitve cells
+    """
+    
+    tm = TM(
+        columnDimensions=(parameters1["sp"]["columnCount"],),
+        cellsPerColumn=parameters1["tm"]["cellsPerColumn"],
+        activationThreshold=parameters1["tm"]["activationThreshold"],
+        initialPermanence=parameters1["tm"]["initialPerm"],
+        connectedPermanence=parameters1["sp"]["synPermConnected"],
+        minThreshold=parameters1["tm"]["minThreshold"],
+        maxNewSynapseCount=parameters1["tm"]["newSynapseCount"],
+        permanenceIncrement=parameters1["tm"]["permanenceInc"],
+        permanenceDecrement=parameters1["tm"]["permanenceDec"],
+        predictedSegmentDecrement=0.0,
+        maxSegmentsPerCell=parameters1["tm"]["maxSegmentsPerCell"],
+        maxSynapsesPerSegment=parameters1["tm"]["maxSynapsesPerSegment"],
+    )
+    
+    activeColumnsA = SDR(parameters1["sp"]["columnCount"])
+    activeColumnsB = SDR(parameters1["sp"]["columnCount"])
+    
+    activeColumnsA.randomize(sparsity=0.4,seed=1)
+    activeColumnsB.randomize(sparsity=0.4,seed=1)
+    
+    # give pattern A - bursting
+    # give pattern B - bursting
+    # give pattern A - should be predicting
+    
+    tm.activateDendrites(True)
+    self.assertTrue(tm.getPredictiveCells().getSum() == 0)
+    predictiveCellsSDR = tm.getPredictiveCells()
+    tm.activateCells(activeColumnsA,True)
+    
+    _print("\nColumnsA")
+    _print("activeCols:"+str(len(activeColumnsA.sparse)))
+    _print("activeCells:"+str(len(tm.getActiveCells().sparse)))
+    _print("predictiveCells:"+str(len(predictiveCellsSDR.sparse)))
+    
+    
+    
+    tm.activateDendrites(True)
+    self.assertTrue(tm.getPredictiveCells().getSum() == 0)
+    predictiveCellsSDR = tm.getPredictiveCells()
+    tm.activateCells(activeColumnsB,True)
+
+    _print("\nColumnsB")
+    _print("activeCols:"+str(len(activeColumnsB.sparse)))
+    _print("activeCells:"+str(len(tm.getActiveCells().sparse)))
+    _print("predictiveCells:"+str(len(predictiveCellsSDR.sparse)))
+    
+    tm.activateDendrites(True)
+    self.assertTrue(tm.getPredictiveCells().getSum() > 0)
+    predictiveCellsSDR = tm.getPredictiveCells()
+    tm.activateCells(activeColumnsA,True)
+    
+    _print("\nColumnsA")
+    _print("activeCols:"+str(len(activeColumnsA.sparse)))
+    _print("activeCells:"+str(len(tm.getActiveCells().sparse)))
+    _print("predictiveCells:"+str(len(predictiveCellsSDR.sparse)))
+    
+
+def _print(txt):
+    if debugPrint:
+        print(txt)
+
+if __name__ == "__main__":
+  unittest.main()
